requirements.txt
smplx[all]
autograd
Cython
easydict
ffmpeg
graphviz
h5py
lmdb
matplotlib
multiprocess
numpy
opencv-python
pathos
Pillow
tqdm
scikit-image
scikit-learn
scipy
sklearn
chumpy
pytorch-lightning==1.3.5
pyvista==0.40.dev0
pyrender
seaborn
protobuf~=3.19.6
progress
imutils
pytube
pyyaml
pynvml
opencv-python
git+https://github.com/mmatl/pyopengl.git#egg=pyopengl
